requirements.txt
# Core Django and DRF
Django==5.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.9

# CORS and Security
django-cors-headers==4.3.1

# Environment and Configuration
python-decouple==3.8

# API Documentation
drf-spectacular==0.27.2

# Email and utilities
Pillow==10.4.0

# Development and Testing
django-extensions==3.2.3

# Production
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0